LayoutTests/fast/events/touch/ios/touch-events-during-scroll-deceleration-in-overflow-expected.txt

PASS document.getElementById('container').scrollTop became different from 0
PASS document.getElementById('containerTouches').textContent is "1"
PASS document.getElementById('documentTouches').textContent is "1"
PASS document.getElementById('containerTouches').textContent is "1"
PASS document.getElementById('documentTouches').textContent is "2"
PASS successfullyParsed is true

TEST COMPLETE
To manually test, swipe down to scroll in the container. Both touch counts should update to 1. Then tap elsewhere in the document. The document touch count should increment to 2, but the container touch count should remain at 1.

The touch count in the scroll container is: 1

The touch count in the document is: 2
